requirements.txt
PyYAML~=6.0.2
tenacity~=9.0.0
pydantic~=2.10.4
selenium~=4.27.1
webdriver-manager~=4.0.2
langchain-openai~=0.2.14
lxml~=5.3.0
beautifulsoup4~=4.12.3
langchain~=0.3.14
langchain-core~=0.3.29
requests~=2.32.3
tqdm~=4.67.1
openai-whisper~=20240930
moviepy~=2.1.1
torch~=2.5.1